move to hooks, use .ts
